refactor: split parser.rs into per-language submodules
Break the 3308-line src/parser.rs into src/parser/ with one module per
language (python, javascript, rust, go, java, csharp, c, cpp) plus a
shared common module for node_text and normalize_type_name helpers. C++
reuses C's call/signature/type/typedef extractors via pub(super).

Public API (parse_source, ParseResult, Symbol, Import, Call) is
unchanged; all 95 integration and 354 unit tests pass without
modification.
